requirements.txt
fal-client==0.5.6
fastapi[all]==0.115.6
python-dotenv==1.0.1